Combine commonly used CSS classes
Previously multiple individual styling classes were being applied to
each search result element in turn. Combining them into new classes
reduces amount of HTML server has to send back.

And in the case of marks in the show email page, this eliminates the
janky JS code that added styling classes at render time. It was done
this way to reduce the amount of string assembly done in the server.